Working on Flushing out Item system
Item is a class that represents an item
used within the  game. ItemActions is
a class that is inteded to be static, and
is used to create the actions of the
items being used. This is a work
around due to the fact that you cannot
assign callbacks to items within the
unity editor . Still trying to figure out
how to assign the action to the item
created

Modified the PlayerController to test
the use of a item which will heal the
character. Can easily refactor that out
later to a more ideal component.